add exercises for tdd
