+++ src/__tests__/tdd-01-markup.finished.js
@@ -4,14 +4,12 @@ import 'react-testing-library/cleanup-after-each'
 
 import React from 'react'
 import {render} from 'react-testing-library'
-import {Editor} from '../post-editor'
+import {Editor} from '../post-editor-01-markup'
 
-test.skip('renders a form with title, content, tags, and a submit button', () => {
+test('renders a form with title, content, tags, and a submit button', () => {
   const {getByLabelText, getByText} = render(<Editor />)
   getByLabelText(/title/i)
   getByLabelText(/content/i)
   getByLabelText(/tags/i)
   getByText(/submit/i)
-  // 🐨 click the submit button
-  // 🐨 verify that the submit button is disabled when clicked
 })
