+++ src/__tests__/tdd-02-state.finished.js
@@ -3,15 +3,17 @@ import 'jest-dom/extend-expect'
 import 'react-testing-library/cleanup-after-each'
 
 import React from 'react'
-import {render} from 'react-testing-library'
-import {Editor} from '../post-editor'
+import {render, fireEvent} from 'react-testing-library'
+import {Editor} from '../post-editor-02-state'
 
-test.skip('renders a form with title, content, tags, and a submit button', () => {
+test('renders a form with title, content, tags, and a submit button', () => {
   const {getByLabelText, getByText} = render(<Editor />)
   getByLabelText(/title/i)
   getByLabelText(/content/i)
   getByLabelText(/tags/i)
-  getByText(/submit/i)
-  // 🐨 click the submit button
-  // 🐨 verify that the submit button is disabled when clicked
+  const submitButton = getByText(/submit/i)
+
+  fireEvent.click(submitButton)
+
+  expect(submitButton).toBeDisabled()
 })
