+++ src/__tests__/tdd-03-api.js
@@ -4,39 +4,31 @@ import 'react-testing-library/cleanup-after-each'
 
 import React from 'react'
 import {render, fireEvent} from 'react-testing-library'
-import {savePost as mockSavePost} from '../api'
-import {Editor} from '../post-editor-03-api'
+// 🐨 you'll need to import your mock savePost from '../api' here
+import {Editor} from '../post-editor'
 
-jest.mock('../api', () => {
-  return {
-    savePost: jest.fn(() => Promise.resolve()),
-  }
-})
+// 🐨 use jest.mock to mock out the ../api module and return a `savePost`
+// jest.fn function that resolves a promise
 
-afterEach(() => {
-  mockSavePost.mockClear()
-})
+// 🐨 after each test, the mock savePost mock function should be cleared (mockClear)
 
-test('renders a form with title, content, tags, and a submit button', () => {
-  const fakeUser = {id: 'user-1'}
-  const {getByLabelText, getByText} = render(<Editor user={fakeUser} />)
-  const fakePost = {
-    title: 'Test Title',
-    content: 'Test content',
-    tags: ['tag1', 'tag2'],
-  }
-  getByLabelText(/title/i).value = fakePost.title
-  getByLabelText(/content/i).value = fakePost.content
-  getByLabelText(/tags/i).value = fakePost.tags.join(', ')
+// 🐨 unskip this test
+test.skip('renders a form with title, content, tags, and a submit button', () => {
+  // 🐨 pass a fake user (an object with an ID) to the editor as a prop
+  const {getByLabelText, getByText} = render(<Editor />)
+
+  // 🐨 set the value of each of these fields
+  getByLabelText(/title/i)
+  getByLabelText(/content/i)
+  // 💯 tags should be a comma-separated list of values here
+  getByLabelText(/tags/i)
   const submitButton = getByText(/submit/i)
 
   fireEvent.click(submitButton)
 
   expect(submitButton).toBeDisabled()
 
-  expect(mockSavePost).toHaveBeenCalledTimes(1)
-  expect(mockSavePost).toHaveBeenCalledWith({
-    ...fakePost,
-    authorId: fakeUser.id,
-  })
+  // 🐨 assert that the mock `savePost` function was called once
+  // and was called with the fake post data (title, content, and tags) and the authorId
+  // 💯 tags should be an array of values here.
 })
