+++ src/__tests__/tdd-06-generate-data.js
@@ -4,10 +4,11 @@ import 'react-testing-library/cleanup-after-each'
 
 import React from 'react'
 import {render, fireEvent, wait} from 'react-testing-library'
-import {build, fake, sequence} from 'test-data-bot'
+// 🐨 you'll need this:
+// import {build, fake, sequence} from 'test-data-bot'
 import {Redirect as MockRedirect} from 'react-router'
 import {savePost as mockSavePost} from '../api'
-import {Editor} from '../post-editor-06-generate-data'
+import {Editor} from '../post-editor'
 
 jest.mock('react-router', () => {
   return {
@@ -26,20 +27,20 @@ afterEach(() => {
   mockSavePost.mockClear()
 })
 
-const postBuilder = build('Post').fields({
-  title: fake(f => f.lorem.words()),
-  content: fake(f => f.lorem.paragraphs().replace(/\r/g, '')),
-  tags: fake(f => [f.lorem.word(), f.lorem.word(), f.lorem.word()]),
-})
+// 🐨 make a postBuilder and a userBuilder using test-data-bot
+// 📖 https://github.com/jackfranklin/test-data-bot
 
-const userBuilder = build('User').fields({
-  id: sequence(s => `user-${s}`),
-})
-
-test('renders a form with title, content, tags, and a submit button', async () => {
-  const fakeUser = userBuilder()
+// 🐨 unskip this test
+test.skip('renders a form with title, content, tags, and a submit button', async () => {
+  // 🐨 swap this object with a call to your userBuilder
+  const fakeUser = {id: 'user-1'}
   const {getByLabelText, getByText} = render(<Editor user={fakeUser} />)
-  const fakePost = postBuilder()
+  // 🐨 swap this object with a call to your postBuilder
+  const fakePost = {
+    title: 'Test Title',
+    content: 'Test content',
+    tags: ['tag1', 'tag2'],
+  }
   const preDate = Date.now()
 
   getByLabelText(/title/i).value = fakePost.title
