+++ src/__tests__/tdd-07-error-state.finished.js
@@ -3,12 +3,11 @@ import 'jest-dom/extend-expect'
 import 'react-testing-library/cleanup-after-each'
 
 import React from 'react'
-// 🐨 you're going to need waitForElement
-import {render, fireEvent, wait} from 'react-testing-library'
+import {render, fireEvent, wait, waitForElement} from 'react-testing-library'
 import {build, fake, sequence} from 'test-data-bot'
 import {Redirect as MockRedirect} from 'react-router'
 import {savePost as mockSavePost} from '../api'
-import {Editor} from '../post-editor'
+import {Editor} from '../post-editor-07-error-state'
 
 jest.mock('react-router', () => {
   return {
@@ -37,8 +36,7 @@ const userBuilder = build('User').fields({
   id: sequence(s => `user-${s}`),
 })
 
-// 🐨 unskip this test
-test.skip('renders a form with title, content, tags, and a submit button', async () => {
+test('renders a form with title, content, tags, and a submit button', async () => {
   const fakeUser = userBuilder()
   const {getByLabelText, getByText} = render(<Editor user={fakeUser} />)
   const fakePost = postBuilder()
@@ -70,10 +68,23 @@ test.skip('renders a form with title, content, tags, and a submit button', async
   expect(MockRedirect).toHaveBeenCalledWith({to: '/'}, {})
 })
 
-// 🐨 add a new test here to verify that it renders a server error.
-// it's very similar to the test above, but it should:
-// at the top: mockSavePost.mockRejectedValueOnce({data: {error: 'some error'}})
-// and at the bottom:
-// assert that an element with the data-testid of 'post-error' appears with the
-// content of the error message
-// and the submitButton is no longer disabled.
+test('renders an error message from the server', async () => {
+  const testError = 'test error'
+  mockSavePost.mockRejectedValueOnce({data: {error: testError}})
+  const fakeUser = userBuilder()
+  const {getByLabelText, getByText, getByTestId} = render(
+    <Editor user={fakeUser} />,
+  )
+  const fakePost = postBuilder()
+
+  getByLabelText(/title/i).value = fakePost.title
+  getByLabelText(/content/i).value = fakePost.content
+  getByLabelText(/tags/i).value = fakePost.tags.join(', ')
+  const submitButton = getByText(/submit/i)
+
+  fireEvent.click(submitButton)
+
+  const postError = await waitForElement(() => getByTestId('post-error'))
+  expect(postError).toHaveTextContent(testError)
+  expect(submitButton).not.toBeDisabled()
+})
