+++ src/__tests__/tdd-08-custom-render.finished.js
@@ -3,12 +3,11 @@ import 'jest-dom/extend-expect'
 import 'react-testing-library/cleanup-after-each'
 
 import React from 'react'
-// 🐨 you're going to need waitForElement
-import {render, fireEvent, wait} from 'react-testing-library'
+import {render, fireEvent, wait, waitForElement} from 'react-testing-library'
 import {build, fake, sequence} from 'test-data-bot'
 import {Redirect as MockRedirect} from 'react-router'
 import {savePost as mockSavePost} from '../api'
-import {Editor} from '../post-editor'
+import {Editor} from '../post-editor-08-custom-render'
 
 jest.mock('react-router', () => {
   return {
@@ -37,17 +36,26 @@ const userBuilder = build('User').fields({
   id: sequence(s => `user-${s}`),
 })
 
-// 🐨 unskip this test
-test.skip('renders a form with title, content, tags, and a submit button', async () => {
+function renderEditor() {
   const fakeUser = userBuilder()
-  const {getByLabelText, getByText} = render(<Editor user={fakeUser} />)
+  const utils = render(<Editor user={fakeUser} />)
   const fakePost = postBuilder()
-  const preDate = Date.now()
 
-  getByLabelText(/title/i).value = fakePost.title
-  getByLabelText(/content/i).value = fakePost.content
-  getByLabelText(/tags/i).value = fakePost.tags.join(', ')
-  const submitButton = getByText(/submit/i)
+  utils.getByLabelText(/title/i).value = fakePost.title
+  utils.getByLabelText(/content/i).value = fakePost.content
+  utils.getByLabelText(/tags/i).value = fakePost.tags.join(', ')
+  const submitButton = utils.getByText(/submit/i)
+  return {
+    ...utils,
+    submitButton,
+    fakeUser,
+    fakePost,
+  }
+}
+
+test('renders a form with title, content, tags, and a submit button', async () => {
+  const {submitButton, fakePost, fakeUser} = renderEditor()
+  const preDate = Date.now()
 
   fireEvent.click(submitButton)
 
@@ -70,10 +78,14 @@ test.skip('renders a form with title, content, tags, and a submit button', async
   expect(MockRedirect).toHaveBeenCalledWith({to: '/'}, {})
 })
 
-// 🐨 add a new test here to verify that it renders a server error.
-// it's very similar to the test above, but it should:
-// at the top: mockSavePost.mockRejectedValueOnce({data: {error: 'some error'}})
-// and at the bottom:
-// assert that an element with the data-testid of 'post-error' appears with the
-// content of the error message
-// and the submitButton is no longer disabled.
+test('renders an error message from the server', async () => {
+  const testError = 'test error'
+  mockSavePost.mockRejectedValueOnce({data: {error: testError}})
+  const {submitButton, getByTestId} = renderEditor()
+
+  fireEvent.click(submitButton)
+
+  const postError = await waitForElement(() => getByTestId('post-error'))
+  expect(postError).toHaveTextContent(testError)
+  expect(submitButton).not.toBeDisabled()
+})
