+++ src/__tests__/tdd-08-custom-render.js
@@ -7,7 +7,7 @@ import {render, fireEvent, wait, waitForElement} from 'react-testing-library'
 import {build, fake, sequence} from 'test-data-bot'
 import {Redirect as MockRedirect} from 'react-router'
 import {savePost as mockSavePost} from '../api'
-import {Editor} from '../post-editor-08-custom-render'
+import {Editor} from '../post-editor'
 
 jest.mock('react-router', () => {
   return {
@@ -36,26 +36,20 @@ const userBuilder = build('User').fields({
   id: sequence(s => `user-${s}`),
 })
 
-function renderEditor() {
+// 🐨 Take the common parts of both tests and create a renderEditor function
+// that does everything those tests both need to do.
+
+// 🐨 unskip this test
+test.skip('renders a form with title, content, tags, and a submit button', async () => {
   const fakeUser = userBuilder()
-  const utils = render(<Editor user={fakeUser} />)
+  const {getByLabelText, getByText} = render(<Editor user={fakeUser} />)
   const fakePost = postBuilder()
-
-  utils.getByLabelText(/title/i).value = fakePost.title
-  utils.getByLabelText(/content/i).value = fakePost.content
-  utils.getByLabelText(/tags/i).value = fakePost.tags.join(', ')
-  const submitButton = utils.getByText(/submit/i)
-  return {
-    ...utils,
-    submitButton,
-    fakeUser,
-    fakePost,
-  }
-}
-
-test('renders a form with title, content, tags, and a submit button', async () => {
-  const {submitButton, fakePost, fakeUser} = renderEditor()
   const preDate = Date.now()
+
+  getByLabelText(/title/i).value = fakePost.title
+  getByLabelText(/content/i).value = fakePost.content
+  getByLabelText(/tags/i).value = fakePost.tags.join(', ')
+  const submitButton = getByText(/submit/i)
 
   fireEvent.click(submitButton)
 
@@ -78,10 +72,20 @@ test('renders a form with title, content, tags, and a submit button', async () =
   expect(MockRedirect).toHaveBeenCalledWith({to: '/'}, {})
 })
 
-test('renders an error message from the server', async () => {
+// 🐨 unskip this test
+test.skip('renders an error message from the server', async () => {
   const testError = 'test error'
   mockSavePost.mockRejectedValueOnce({data: {error: testError}})
-  const {submitButton, getByTestId} = renderEditor()
+  const fakeUser = userBuilder()
+  const {getByLabelText, getByText, getByTestId} = render(
+    <Editor user={fakeUser} />,
+  )
+  const fakePost = postBuilder()
+
+  getByLabelText(/title/i).value = fakePost.title
+  getByLabelText(/content/i).value = fakePost.content
+  getByLabelText(/tags/i).value = fakePost.tags.join(', ')
+  const submitButton = getByText(/submit/i)
 
   fireEvent.click(submitButton)
